Add kill-switch
- Add a kill-switch to disable all token retrieval.
- Make `GitHubTokenBroker` a singleton.
