+++ src/Costellobot/appsettings.json
@@ -66,6 +66,7 @@
     },
     "Scopes": [],
     "TokenBroker": {
+      "IsEnabled": true,
       "Repositories": {
         "martincostello/adventofcode": {
           "actions": {
